'moved stuff into lib and 3rdparty'
