+++ lf/system/bootstrap.php
@@ -9,14 +9,13 @@
 define('ENV', 'PRODUCTION');
 
 // Littlefoot
-require 'system/functions.php'; // Helpful functions
-require 'system/db.class.php'; // Legacy Database Wrapper
+require 'system/lib/helpers.php'; // Helpful functions
 require 'system/lib/db.php'; // OOP Database Wrapper
-require 'system/app.class.php'; // Littlefoot app base class
+require 'system/lib/app.php'; // Littlefoot app base class
 require 'system/lib/orm.php'; // Object Relation Model base
 require 'system/lib/recovery/install.php';
-require 'system/littlefoot.php'; // Littlefoot CMS (Request, Auth, Run assigned Apps, Render on template)
 require 'system/lib/auth.php'; // auth stuff
+require 'system/lib/littlefoot.php'; // Request, Auth, Nav, Content, Render
 
 // Add local lib paths to include_path
 if(is_dir(LF.'lib'))
