+++ lf/system/lib/3rdparty/icons/FontAwesome.otf
