+++ lf/system/lib/3rdparty/icons/fontawesome-webfont.ttf
