+++ lf/system/lib/helpers.php
